+++ src/networkx_temporal/algorithms/community/spectral/bethe_hessian.py
@@ -3,10 +3,10 @@ import scipy.sparse as sp
 from typing import List, Optional, Union
 from warnings import warn
 
-from ...cugraph import NX_CUGRAPH_AUTOCONFIG
+from ...gpu import NX_GPU_AUTOCONFIG
 from ....typing import Literal
 
-DEVICE = "gpu" if NX_CUGRAPH_AUTOCONFIG else "cpu"
+DEVICE = "gpu" if NX_GPU_AUTOCONFIG else "cpu"
 
 
 def spectral_clustering_bethe_hessian(
@@ -49,7 +49,7 @@ def spectral_clustering_bethe_hessian(
 
     .. hint::
 
-       Setting ``NX_CUGRAPH_AUTOCONFIG=1`` in the environment will set ``device='gpu'`` as default.
+       Setting ``NX_GPU_AUTOCONFIG=1`` in the environment will set ``device='gpu'`` as default.
 
     .. seealso::