+++ src/networkx_temporal/algorithms/gpu.py
@@ -1,5 +1,9 @@
 from os import environ
 
+NX_GPU_AUTOCONFIG = False
+if environ.get("NX_GPU_AUTOCONFIG", "0").lower() in ("1", "true"):
+    NX_GPU_AUTOCONFIG = True
+
 NX_CUGRAPH_AUTOCONFIG = False
 if environ.get("NX_CUGRAPH_AUTOCONFIG", "0").lower() in ("1", "true"):
     NX_CUGRAPH_AUTOCONFIG = True